# docs/swagger.yaml
basePath: /
definitions:
...
  databaseTypes.SchoolStore:
    properties:
      Category:
        example: 2
        type: integer
      Date_Added:
        example: "2022-01-01T12:00:00Z"
        type: string
      Description:
        type: string
      ID:
        example: 1
        type: integer
      Price:
        example: 49.99
        type: number
      Product_Name:
        example: Backpack
        type: string
      Stock:
        example: 10
        type: integer
      image_file:
        items:
          type: integer
        type: array
    type: object
...
  restTypes.ErrorResponse:
    properties:
      code:
        description: |-
          HTTP status code of the error response.

          Example: 400

          Required: true
        type: integer
      message:
        description: |-
          Error message.

          Example: Invalid request

          Required: true
        type: string
    type: object
...
  restTypes.SchoolStorePostResponse:
    properties:
      id:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
  restTypes.SchoolStoreResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/databaseTypes.SchoolStore'
        type: array
    type: object
...
info:
  contact:
    name: Senya
  description: Simple swagger implementation in Go HTTP
  title: Go Rest API with Swagger for school system
  version: "1.0"
paths:
...
  /data/school-store/:
    get:
      consumes:
      - application/json
      description: Retrieves a list of items from the School Store database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/restTypes.SchoolStoreResponse'
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get a list of items from the School Store
      tags:
      - School Store
    post:
      consumes:
      - multipart/form-data
      description: Adds a new item to the School Store database
      parameters:
      - description: Name of the item to add
        in: formData
        name: item_name
        required: true
        type: string
      - description: Description of the item to add
        in: formData
        name: description
        required: true
        type: string
      - description: Price of the item to add
        in: formData
        name: price
        required: true
        type: number
      - description: Category of the item to add
        in: formData
        name: category
        required: true
        type: string
      - description: Image file of the item to add
        in: formData
        name: image_file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/restTypes.ErrorResponse'
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Add an item to the School Store
      tags:
      - School Store
  /data/school-store/{item_id}:
    delete:
      description: Deletes an item from the School Store database
      parameters:
      - description: ID of the item to delete
        in: path
        name: item_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/restTypes.ErrorResponse'
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "404":
          description: Item not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Delete an item from the School Store
      tags:
      - School Store
    put:
      consumes:
      - multipart/form-data
      description: Updates an existing item in the School Store database
      parameters:
      - description: ID of the item to update
        in: path
        name: item_id
        required: true
        type: integer
      - description: New name of the item
        in: formData
        name: item_name
        type: string
      - description: New description of the item
        in: formData
        name: description
        type: string
      - description: New price of the item
        in: formData
        name: price
        type: number
      - description: New category of the item
        in: formData
        name: category
        type: string
      - description: New image file of the item
        in: formData
        name: image_file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/restTypes.SchoolStorePostResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Bearer: []
      summary: Update an item in the School Store
      tags:
      - School Store
  /data/school-store/image/{item_id}:
    get:
      consumes:
      - application/json
      description: Retrieves an image for a specified item from the School Store database
      parameters:
      - description: ID of the item to retrieve the image for
        in: path
        name: item_id
        required: true
        type: integer
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Invalid request parameters
          schema:
            type: string
        "404":
          description: Item not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get an image for an item from the School Store
      tags:
      - School Store
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"